Move worker functionality in package
